+++ src/NtUtil.h
@@ -1,0 +1,5 @@
+#pragma once
+
+LPWSTR GetObjectName(HANDLE hObject);
+
+LPWSTR GetObjectTypeName(HANDLE hObject);
